+++ CHANGELOG.md
@@ -5,8 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.5.0] - 2024-09-05
 ### Added
 - Improved warning log for conditional visibility without default
+- Datetime and time in table field
 
 ### Changed
 - Pinned dmc >= 0.14.4 for datetime bugfix